requirements.txt
Django==3.0.2
django-cors-headers==3.2.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
elasticsearch==7.1.0
gunicorn==20.0.4